+++ site/sidebar.ts
@@ -13,33 +13,75 @@ export const sidebar = {
       ],
     },
     {
-      text: 'App',
+      text: 'SDK',
       collapsed: true,
-      items: [],
+      items: [
+        { text: 'Installation', link: '/sdk/installation' },
+        { text: 'Quote', link: '/sdk/quote' },
+        { text: 'Swap', link: '/sdk/swap' },
+        // { text: 'Limit', link: '/sdk/limit' },
+        // { text: 'DCA', link: '/sdk/dca' },
+        // { text: 'Cross-Chain Swap', link: '/sdk/cross-chain-swap' },
+        { text: 'Viem', link: '/sdk/viem' },
+      ],
     },
     {
       text: 'API',
       collapsed: true,
       items: [
-        { text: 'Price', link: '/api/price' },
-        { text: 'Quote', link: '/api/quote' },
         { text: 'Swap', link: '/api/swap' },
+        { text: 'Price', link: '/api/price' },
+        // { text: 'Quote', link: '/api/quote' },
       ],
     },
+    // {
+    //   text: 'Widget',
+    //   collapsed: true,
+    //   items: [],
+    // },
+    // {
+    //   text: 'App',
+    //   collapsed: true,
+    //   items: [
+    //     { text: 'Swap', link: '/app/swap' },
+    //     { text: 'Limit', link: '/app/limit' },
+    //     { text: 'DCA', link: '/app/dca' },
+    //     { text: 'Cross-Chain Swap', link: '/app/cross-chain-swap' },
+    //   ],
+    // },
     {
-      text: 'Package',
+      text: 'AMM',
       collapsed: true,
-      items: [],
-    },
-    {
-      text: 'Widget',
-      collapsed: true,
-      items: [],
+      items: [
+        {
+          text: 'clAMM',
+          link: '/amm/clamm',
+        },
+        {
+          text: 'cpAMM',
+          link: '/amm/cpamm',
+        },
+        {
+          text: 'lbAMM',
+          link: '/amm/lbamm',
+        },
+        {
+          text: 'fAMM',
+          link: '/amm/famm',
+        },
+      ],
     },
     {
       text: 'Contracts',
       collapsed: true,
-      items: [],
+      items: [
+        { text: 'SUSHI', link: '/contracts/sushi' },
+        { text: 'xSUSHI', link: '/contracts/xsushi' },
+        { text: 'Route Processor', link: '/contracts/route-processor' },
+        { text: 'clAMM', link: '/contracts/clamm' },
+        { text: 'cpAMM', link: '/contracts/cpamm' },
+        // { text: 'lbAMM', link: '/contracts/lbamm' },
+      ],
     },
     // {
     //   text: 'Guides',