+++ src/config/token-maps/bases-to-check-trades-against.ts
@@ -1,5 +1,13 @@
 import { ChainId } from '../../chain/index.js'
-import { METH, MUSD, SKL, Token, USDB, USDP } from '../../currency/index.js'
+import {
+  METH,
+  MUSD,
+  SKL,
+  STONE,
+  Token,
+  USDB,
+  USDP,
+} from '../../currency/index.js'
 import {
   AAVE,
   BUSD,
@@ -526,5 +534,28 @@ export const BASES_TO_CHECK_TRADES_AGAINST: {
     WNATIVE[ChainId.CURTIS],
     USDT[ChainId.CURTIS],
     DAI[ChainId.CURTIS],
+  ],
+  [ChainId.MANTA]: [
+    WNATIVE[ChainId.MANTA],
+    USDC[ChainId.MANTA],
+    USDT[ChainId.MANTA],
+    STONE[ChainId.MANTA],
+  ],
+  [ChainId.MODE]: [
+    WNATIVE[ChainId.MODE],
+    USDC[ChainId.MODE],
+    USDT[ChainId.MODE],
+  ],
+  [ChainId.TAIKO]: [
+    WNATIVE[ChainId.TAIKO],
+    USDC[ChainId.TAIKO],
+    USDT[ChainId.TAIKO],
+    DAI[ChainId.TAIKO],
+  ],
+  [ChainId.ZKLINK]: [
+    WNATIVE[ChainId.ZKLINK],
+    USDC[ChainId.ZKLINK],
+    USDT[ChainId.ZKLINK],
+    DAI[ChainId.ZKLINK],
   ],
 }